+++ page-objects/pages/NWSNewsListPage.js
@@ -12,9 +12,13 @@ async function getListOfNewNames () {
 	return elementHelpers.getTextFromAllElements(newNameSel);
 }
 
-async function isNewsVisible (name, expectedValue) {
-	let newsNames = await getListOfNewNames();
-	expectedValue ? await expect(newsNames).to.include(name) : await expect(newsNames).to.not.include(name);
+async function isNewsVisible (newsTitle, expectedValue) {
+	const newsTitles = await getListOfNewNames();
+	const isNewskOnList = newsTitles.some((element) => element.includes(newsTitle));
+	const fillString = !expectedValue ? 'not' : '';
+	const msg = 'News with name '+ fillString + ' is visible on the list: \n';
+	const resultMsg = 'Expected: ' + newsTitle + ', Actual: ' + newsTitles;
+	await expect(isNewskOnList, msg + resultMsg).to.equal(expectedValue)
 }
 
 module.exports = {