+++ scripts/ci/end-to-end-tests.travis.sh
@@ -42,6 +42,13 @@ fetch(){
 
 install(){
 	cd docker-compose
+
+	# add -e on mac, use ; as alternative separator
+	sed -i "s/ES_USER.*/ES_USER=${ES_USER}/" docker-compose.end-to-end-tests.yml
+	sed -i "s/ES_PASSWORD.*/ES_PASSWORD=${ES_PASSWORD}/" docker-compose.end-to-end-tests.yml
+	sed -i "s/SECRET_ES_MERLIN_USERNAME.*/SECRET_ES_MERLIN_USERNAME=${SECRET_ES_MERLIN_USERNAME}/" docker-compose.end-to-end-tests.yml
+	sed -i "s/SECRET_ES_MERLIN_PW.*/SECRET_ES_MERLIN_PW=${SECRET_ES_MERLIN_PW}/" docker-compose.end-to-end-tests.yml
+
 	docker-compose -f docker-compose.end-to-end-tests.yml build --parallel
 	docker-compose -f docker-compose.end-to-end-tests.yml up -d
 	cd ..