+++ step_definitions/administration-steps.js
@@ -2,23 +2,71 @@
 
 const ADMNSTRTNAdministrationOverviewPage = require('../page-objects/pages/administrationPages/ADMNSTRTNAdministrationOverviewPage');
 const studentAdministration = require('../page-objects/pages/administrationPages/ADMNSTRTNAdministerStudentsPage');
+const studentEditPage = require('../page-objects/pages/administrationPages/ADMNSTRTNEditDeleteStudentPage');
 
-When(/^admin goes to students administration$/, function () {
-	return ADMNSTRTNAdministrationOverviewPage.clickAdministrateStudents();
+When(/^.* goes to students administration$/, function () {
+	return ADMNSTRTNAdministrationOverviewPage.goToAdministrateStudents();
 });
 
 When(/^.*set student firstname '([^']*)', lastname '([^']*)', email '([^']*)'$/, function (firstname, lastname, email) {
 	return studentAdministration.createNewPupil(firstname, lastname, email);
 });
 
-Then(/^the admin should see new pupil with email '([^']*)' among his pupils$/, function (email) {
+Then(/^.*student with email '([^']*)' is visible on the list$/, function (email) {
 	return studentAdministration.isStudentEmailOnTheList(email);
 });
 
-Then(/^.* manually submits a consent '([^']*)'$/, function (email) {
+When(/^.*manually submits consent for user with e-mail '([^']*)', thus generates a random password for him$/, function (
+	email
+) {
 	return studentAdministration.submitConsent(email);
 });
 
-Then(/^new pupil '([^']*)' can log in with default password$/, async function (email) {
-	await studentAdministration.studentLogsInWithDefaultPassword(email);
+/**  Edit a student */
+When(/^.* clicks Edit-student button$/, async function () {
+	await studentAdministration.clickEditStudentBtn();
+});
+
+When(/^.* clicks edit student Cancel button$/, async function () {
+	await studentEditPage.clickCancelButton();
+});
+
+When(/^.* clicks edit student Cancel button inside popup$/, async function () {
+	await studentEditPage.clickCancelInModal();
+});
+
+When(/^.* clicks edit student Discard-change button inside popup$/, async function () {
+	await studentEditPage.clickDiscardChangesInModal();
+});
+
+When(/^.* changes student firstname to '([^']*)'$/, async function (firstName) {
+	await studentEditPage.setStudentFirstName(firstName);
+});
+
+When(/^.* changes student lastname to '([^']*)'$/, async function (lastName) {
+	await studentEditPage.setStudentLastName(lastName);
+});
+
+When(/^.* changes student email to '([^']*)'$/, async function (email) {
+	await studentEditPage.setStudentEmail(email);
+});
+
+When(/^.* changes student birthdate to '([^']*)'$/, async function (birthdate) {
+	await studentEditPage.setStudentBirthday(birthdate);
+});
+
+Then(/^.* student firstname '([^']*)' is visible on the list$/, async function (firstname) {
+	await studentAdministration.isStudentFirstnameOnTheList(firstname);
+});
+
+Then(/^.* student lastname '([^']*)' is visible on the list$/, async function (lastname) {
+	await studentAdministration.isStudentLastnameOnTheList(lastname);
+});
+
+Then(/^.* student email '([^']*)' is is visible on the list$/, async function (emailAddress) {
+	await studentAdministration.isStudentEmailOnTheList(emailAddress);
+});
+
+Then(/^.* student birthdate is '([^']*)'$/, async function (birthdate) {
+	await studentEditPage.isStudentBirthdayCorrect(birthdate);
 });